+++ root/app/ondemand/shared_state.py
@@ -1,0 +1,4 @@
+import threading
+
+last_accessed_urls = set()
+last_accessed_urls_lock = threading.Lock()
